DataAcquisition/requirements.txt
requests~=2.32.4
yfinance~=0.2.56
fastapi~=0.115.8
pydantic~=2.10.6
python-dotenv~=1.1.0
alpaca-py
uvicorn[standard]==0.23.1
requests~=2.32.4
pytest~=7.3.2
httpx~=0.24.0
